move context tests back together with context
